calcMaxChildren for MCTS shifted into Node, paperplayers moved into aiplayers
