+++ src/main/java/risk/aiGameCoordinator/GlickoGames.java
@@ -16,13 +16,13 @@ import risk.aiplayers.MCTSPlayers.MCTSFull_Baseline_AI;
 import risk.aiplayers.MCTSPlayers.MCTSGenerate_Low_Children_AI;
 import risk.aiplayers.MCTSPlayers.MCTSHashing_Uniqueness_AI;
 import risk.aiplayers.MCTSPlayers.MCTSSample_Duplic_AI;
-import risk.paperplayers.S_BaselineAI;
-import risk.paperplayers.S_EMM_AI;
-import risk.paperplayers.S_MCTS_AI;
-import risk.paperplayers.S_MCTS_Baseline_AI;
-import risk.paperplayers.S_MCTS_Explore_AI;
-import risk.paperplayers.S_MCTS_Naive_AI;
-import risk.paperplayers.S_Simulation_AI;
+import risk.aiplayers.paperplayers.S_BaselineAI;
+import risk.aiplayers.paperplayers.S_EMM_AI;
+import risk.aiplayers.paperplayers.S_MCTS_AI;
+import risk.aiplayers.paperplayers.S_MCTS_Baseline_AI;
+import risk.aiplayers.paperplayers.S_MCTS_Explore_AI;
+import risk.aiplayers.paperplayers.S_MCTS_Naive_AI;
+import risk.aiplayers.paperplayers.S_Simulation_AI;
 import Glicko2.Rating;
 import Glicko2.RatingCalculator;
 import Glicko2.RatingPeriodResults;
