+++ src/main/java/risk/aiplayers/MCTSPlayers/MCTSFairExpansion_AI.java
@@ -30,124 +30,6 @@ public class MCTSFairExpansion_AI extends MonteCarloTreeSearchPlayer {
 			long time) {
 		super(name, opp, map, id, time, new AIParameter());
 	}
-
-	@Override
-	protected void calculateMaxChildren(MCTSNode lastNode) {
-		int count = 0;
-		switch (lastNode.getTreePhase()) {
-		case GameTreeNode.RECRUIT: {
-			int troops = AIUtil.calculateRecruitedTroops(lastNode);
-
-			int n = troops;
-			int m = lastNode.getGame().getCurrentPlayer().getTerritories()
-					.size();
-			count = (int) (AIUtil.nCk(n + m - 1, n));
-
-			if (count < 0)
-				count = Integer.MAX_VALUE;
-
-			break;
-		}
-		case GameTreeNode.ATTACK: {
-			count = 1;
-			lastNode.noAttackAdded = false;
-			Iterator<Territory> it = lastNode.getGame().getCurrentPlayer()
-					.getTerritories().values().iterator();
-			while (it.hasNext()) {
-				Territory t = it.next();
-				// Only consider fortified territories
-				if (t.getNrTroops() > 1) {
-					for (Territory n : t.getNeighbours()) {
-						Territory tempT = lastNode.getGame().getOtherPlayer()
-								.getTerritoryByName(n.getName());
-						// Only consider targets that have less troops than the
-						// source
-						if (tempT != null) {
-							count++;
-						}
-					}
-				}
-			}
-
-			break;
-		}
-		case GameTreeNode.MOVEAFTERATTACK: {
-			count = lastNode.getGame().getCurrentPlayer()
-					.getTerritoryByName(lastNode.getParent().getAttackSource())
-					.getNrTroops() - 1;
-			if (count == 0) {
-				System.out.println("WTF");
-			}
-			break;
-		}
-		case GameTreeNode.RANDOMEVENT: {
-			int sourceTroops = lastNode.getGame().getCurrentPlayer()
-					.getTerritoryByName(lastNode.getAttackSource())
-					.getNrTroops();
-			int destTroops = lastNode.getGame().getOtherPlayer()
-					.getTerritoryByName(lastNode.getAttackDest()).getNrTroops();
-
-			if (sourceTroops == 2) {
-				count = 2;
-			} else if (sourceTroops == 3) {
-				if (destTroops == 2 || destTroops == 1) {
-					count = 2;
-				} else {
-					count = 3;
-				}
-
-			} else {
-				if (destTroops == 2 || destTroops == 1) {
-					count = 2;
-				} else {
-					count = 3;
-				}
-			}
-			break;
-		}
-		case GameTreeNode.MANOEUVRE: {
-			int size = AIUtil.updateRegions(lastNode.getGame());
-			// Create list of connected components
-			lastNode.setConnComponentBuckets(new LinkedList<LinkedList<Territory>>());
-			for (int i = 0; i < size; i++) {
-				lastNode.getConnComponentBuckets().add(
-						new LinkedList<Territory>());
-			}
-
-			// Since not maneuvering is an option
-			count = 1;
-
-			Iterator<Territory> it = lastNode.getGame().getCurrentPlayer()
-					.getTerritories().values().iterator();
-			while (it.hasNext()) {
-				Territory t = it.next();
-				lastNode.getConnComponentBuckets().get(t.connectedRegion)
-						.add(t);
-			}
-
-			for (LinkedList<Territory> bucket : lastNode
-					.getConnComponentBuckets()) {
-				if (bucket.size() > 1) {
-					for (Territory src : bucket) {
-						if (src.getNrTroops() > 1) {
-							for (Territory dest : bucket) {
-								if (!src.getName().equals(dest.getName())) {
-									// Unique source-dest combo
-									count += (src.getNrTroops() - 1);
-								}
-							}
-						}
-					}
-				}
-			}
-			
-			break;
-		}
-		}
-
-		lastNode.maxChildren = count;
-	}
-
 	
 	// EXPANSION (One child at a time)
 	@Override
@@ -211,8 +93,6 @@ public class MCTSFairExpansion_AI extends MonteCarloTreeSearchPlayer {
 
 				maxChild.setMoveReq(false);
 
-				calculateMaxChildren(maxChild);
-
 				maxChild.depth = lastNode.depth + 1;
 				if (maxChild.depth > maxTreeDepth) {
 					maxTreeDepth = maxChild.depth;
@@ -226,7 +106,7 @@ public class MCTSFairExpansion_AI extends MonteCarloTreeSearchPlayer {
 			if (lastNode.numberOfAttackBranches == 0) {
 				lastNode.numberOfAttackBranches = Math.min(
 						params.MCTSAttackBranchQualityFactor,
-						lastNode.maxChildren / 2);
+						lastNode.maxChildren() / 2);
 			}
 
 			// Generate all and pick randomly from top 30
@@ -320,7 +200,6 @@ public class MCTSFairExpansion_AI extends MonteCarloTreeSearchPlayer {
 							lastNode.noAttackAdded = true;
 							maxChild.setTreePhase(GameTreeNode.MANOEUVRE);
 						}
-						calculateMaxChildren(maxChild);
 
 						maxChild.depth = lastNode.depth + 1;
 						if (maxChild.depth > maxTreeDepth) {
@@ -368,7 +247,6 @@ public class MCTSFairExpansion_AI extends MonteCarloTreeSearchPlayer {
 					if (maxChild.getAttackDest().length() == 0) {
 						maxChild.setTreePhase(GameTreeNode.MANOEUVRE);
 					}
-					calculateMaxChildren(maxChild);
 
 					maxChild.depth = lastNode.depth + 1;
 					if (maxChild.depth > maxTreeDepth) {
@@ -480,7 +358,6 @@ public class MCTSFairExpansion_AI extends MonteCarloTreeSearchPlayer {
 			} else {
 				newChild.setTreePhase(GameTreeNode.ATTACK);
 			}
-			calculateMaxChildren(newChild);
 
 			getValue(newChild, lastNode);
 			newChild.depth = lastNode.depth + 1;
@@ -534,7 +411,6 @@ public class MCTSFairExpansion_AI extends MonteCarloTreeSearchPlayer {
 
 					newChild.setMoveReq(false);
 
-					calculateMaxChildren(newChild);
 					getValue(newChild, lastNode);
 					newChild.depth = lastNode.depth + 1;
 					if (newChild.depth > maxTreeDepth) {
@@ -554,7 +430,7 @@ public class MCTSFairExpansion_AI extends MonteCarloTreeSearchPlayer {
 			if (lastNode.numberOfManoeuvreBranches == 0) {
 				lastNode.numberOfManoeuvreBranches = Math.min(
 						params.MCTSManBranchQualityFactor,
-						lastNode.maxChildren / 4);
+						lastNode.maxChildren() / 4);
 			}
 
 			if(lastNode.manSources == null) {
@@ -603,7 +479,7 @@ public class MCTSFairExpansion_AI extends MonteCarloTreeSearchPlayer {
 				double maxRating = Double.NEGATIVE_INFINITY;
 				MCTSNode maxChild = null;
 
-				if (lastNode.maxChildren == 1) {
+				if (lastNode.maxChildren() == 1) {
 					maxChild = lastNode.clone();
 					maxChild.setTreePhase(GameTreeNode.RECRUIT);
 					maxChild.switchMaxPlayer();
@@ -612,7 +488,7 @@ public class MCTSFairExpansion_AI extends MonteCarloTreeSearchPlayer {
 				}
 
 				// Fix search range
-				if ((lastNode.maxChildren - lastNode.getChildren().size()) < lastNode.numberOfManoeuvreBranches + 1) {
+				if ((lastNode.maxChildren() - lastNode.getChildren().size()) < lastNode.numberOfManoeuvreBranches + 1) {
 					lastNode.numberOfManoeuvreBranches--;
 				}
 
@@ -621,7 +497,7 @@ public class MCTSFairExpansion_AI extends MonteCarloTreeSearchPlayer {
 				}
 
 				for (int i = 0; i < lastNode.numberOfManoeuvreBranches; i++) {
-					int index = r.nextInt(lastNode.maxChildren);
+					int index = r.nextInt(lastNode.maxChildren());
 
 					int first, last, middle = -1, nrTroops = -1;
 					MCTSNode temp = null;
@@ -690,7 +566,6 @@ public class MCTSFairExpansion_AI extends MonteCarloTreeSearchPlayer {
 						maxChild.getGame().changeCurrentPlayer();
 						
 						maxChild.setMoveReq(false);
-						calculateMaxChildren(maxChild);
 
 						getValue(maxChild, lastNode);
 						maxChild.depth = lastNode.depth + 1;
@@ -745,8 +620,6 @@ public class MCTSFairExpansion_AI extends MonteCarloTreeSearchPlayer {
 
 					maxChild.setMoveReq(false);
 
-					calculateMaxChildren(maxChild);
-					
 					maxChild.depth = lastNode.depth + 1;
 					if (maxChild.depth > maxTreeDepth) {
 						maxTreeDepth = maxChild.depth;
@@ -784,7 +657,7 @@ public class MCTSFairExpansion_AI extends MonteCarloTreeSearchPlayer {
 
 			int n = numberOfTroops;
 			int m = game.getCurrentPlayer().getTerritories().size();
-			root.maxChildren = (int) (AIUtil.nCk(n + m - 1, n));
+			root.setMaxChildren((int) (AIUtil.nCk(n + m - 1, n)));
 
 			treeDepth = 0;
 			maxTreeDepth = Integer.MIN_VALUE;
@@ -998,7 +871,6 @@ public class MCTSFairExpansion_AI extends MonteCarloTreeSearchPlayer {
 				globalNode.setWinCount(0);
 				globalNode.depth = 1;
 				globalNode.setChildren(new ArrayList<MCTSNode>());
-				calculateMaxChildren(globalNode);
 			}
 		}
