+++ src/main/java/risk/aiplayers/MCTSPlayers/MCTSFull_Baseline_AI.java
@@ -32,7 +32,7 @@ public class MCTSFull_Baseline_AI extends MonteCarloTreeSearchPlayer {
 			long time) {
 		super(name, opp, map, id, time, new AIParameter());
 	}
-	@Override
+
 	protected void calculateMaxChildren(MCTSNode lastNode) {
 		int count = 0;
 		switch (lastNode.getTreePhase()) {
@@ -133,7 +133,7 @@ public class MCTSFull_Baseline_AI extends MonteCarloTreeSearchPlayer {
 			break;
 		}
 		}
-		lastNode.maxChildren = count;
+		lastNode.setMaxChildren(count);
 	}
 	// EXPANSION (One child at a time)
 	@Override
@@ -227,7 +227,7 @@ public class MCTSFull_Baseline_AI extends MonteCarloTreeSearchPlayer {
 			if (lastNode.numberOfAttackBranches == 0) {
 				lastNode.numberOfAttackBranches = Math.min(
 						params.MCTSAttackBranchQualityFactor,
-						lastNode.maxChildren / 2);
+						lastNode.maxChildren() / 2);
 			}
 			// Generate all and pick randomly from top 30
 			if (lastNode.attackChildren == null) {
@@ -242,7 +242,7 @@ public class MCTSFull_Baseline_AI extends MonteCarloTreeSearchPlayer {
 					// Add option to not attack
 					lastNode.attackChildren.add(noAttackChild);
 					// Decrease Max Children then
-					lastNode.maxChildren--;
+					lastNode.setMaxChildren(lastNode.maxChildren()-1);
 				}
 				// Populate treeset
 				Iterator<Territory> it = lastNode.getGame().getCurrentPlayer()
@@ -286,7 +286,7 @@ public class MCTSFull_Baseline_AI extends MonteCarloTreeSearchPlayer {
 					lastNode.numberOfAttackBranches--;
 				}
 				if (lastNode.numberOfAttackBranches <= 1) {
-					lastNode.numberOfAttackBranches = lastNode.maxChildren - lastNode.getChildren().size();
+					lastNode.numberOfAttackBranches = lastNode.maxChildren() - lastNode.getChildren().size();
 				}
 				MCTSNode maxChild = null;
 				double maxRating = Double.NEGATIVE_INFINITY;
@@ -395,7 +395,7 @@ public class MCTSFull_Baseline_AI extends MonteCarloTreeSearchPlayer {
 					}
 				}
 				// Catching forever loop
-				if (lastNode.numberOfAttackBranches <= 1 && (lastNode.maxChildren - lastNode.getChildren().size()) == 1) {
+				if (lastNode.numberOfAttackBranches <= 1 && (lastNode.maxChildren() - lastNode.getChildren().size()) == 1) {
 					for (MCTSNode child : lastNode.getChildren()) {
 						if (child.getAttackSource().equalsIgnoreCase(
 								maxChild.getAttackSource())
@@ -605,7 +605,7 @@ public class MCTSFull_Baseline_AI extends MonteCarloTreeSearchPlayer {
 			if (lastNode.numberOfManoeuvreBranches == 0) {
 				lastNode.numberOfManoeuvreBranches = Math.min(
 						params.MCTSManBranchQualityFactor,
-						lastNode.maxChildren / 4);
+						lastNode.maxChildren() / 4);
 			}
 			if(lastNode.manSources == null) {
 				lastNode.manSources = new ArrayList<Territory>();
@@ -648,7 +648,7 @@ public class MCTSFull_Baseline_AI extends MonteCarloTreeSearchPlayer {
 				double maxRating = Double.NEGATIVE_INFINITY;
 				MCTSNode maxChild = null;
 				
-				if (lastNode.maxChildren == 1) {
+				if (lastNode.maxChildren() == 1) {
 					maxChild = lastNode.clone();
 					maxChild.setTreePhase(GameTreeNode.RECRUIT);
 					maxChild.switchMaxPlayer();
@@ -656,7 +656,7 @@ public class MCTSFull_Baseline_AI extends MonteCarloTreeSearchPlayer {
 					maxChild.updateHash(lastNode);
 				}
 				// Fix search range
-				if ((lastNode.maxChildren - lastNode.getChildren().size()) < lastNode.numberOfManoeuvreBranches + 1) {
+				if ((lastNode.maxChildren() - lastNode.getChildren().size()) < lastNode.numberOfManoeuvreBranches + 1) {
 					lastNode.numberOfManoeuvreBranches--;
 				}
 				if (lastNode.numberOfManoeuvreBranches < 1) {
@@ -718,7 +718,7 @@ public class MCTSFull_Baseline_AI extends MonteCarloTreeSearchPlayer {
 					}
 				} else {
 					for (int i = 0; i < lastNode.numberOfManoeuvreBranches; i++) {
-						int index = r.nextInt(lastNode.maxChildren);
+						int index = r.nextInt(lastNode.maxChildren());
 						int first, last, middle = -1, nrTroops = -1;
 						MCTSNode temp = null;
 						if (index > 0) {
@@ -870,7 +870,7 @@ public class MCTSFull_Baseline_AI extends MonteCarloTreeSearchPlayer {
 		root.setChildren(new ArrayList<MCTSNode>());
 		int n = numberOfTroops;
 		int m = game.getCurrentPlayer().getTerritories().size();
-		root.maxChildren = (int) (AIUtil.nCk(n + m - 1, n));
+		root.setMaxChildren((int) (AIUtil.nCk(n + m - 1, n)));
 		treeDepth = 0;
 		maxTreeDepth = Integer.MIN_VALUE;
 		treeNodeCount = 0;
