+++ src/main/java/risk/aiplayers/MCTSPlayers/MCTSMomentum_AI.java
@@ -88,8 +88,6 @@ public class MCTSMomentum_AI extends MCTSMove_After_Attack_AI{
 
 			maxChild.setMoveReq(false);
 
-			calculateMaxChildren(maxChild);
-
 			maxChild.depth = lastNode.depth + 1;
 			if (maxChild.depth > maxTreeDepth) {
 				maxTreeDepth = maxChild.depth;
@@ -103,7 +101,7 @@ public class MCTSMomentum_AI extends MCTSMove_After_Attack_AI{
 			if (lastNode.numberOfAttackBranches == 0) {
 				lastNode.numberOfAttackBranches = Math.min(
 						params.MCTSAttackBranchQualityFactor,
-						(lastNode.maxChildren - lastNode.getChildren().size()) / 2);
+						(lastNode.maxChildren() - lastNode.getChildren().size()) / 2);
 			}
 
 			// Generate all and pick randomly from top 30
@@ -158,8 +156,6 @@ public class MCTSMomentum_AI extends MCTSMove_After_Attack_AI{
 					MomentumChild.setWinCount(0);
 					MomentumChild.setParent(lastNode);
 					MomentumChild.setChildren(new ArrayList<MCTSNode>());
-
-					calculateMaxChildren(MomentumChild);
 
 					MomentumChild.depth = lastNode.depth + 1;
 					if (MomentumChild.depth > maxTreeDepth) {
@@ -222,7 +218,6 @@ public class MCTSMomentum_AI extends MCTSMove_After_Attack_AI{
 							lastNode.noAttackAdded = true;
 							maxChild.setTreePhase(GameTreeNode.MANOEUVRE);
 						}
-						calculateMaxChildren(maxChild);
 						maxChild.updateHash(lastNode);
 						
 						maxChild.depth = lastNode.depth + 1;
@@ -242,7 +237,6 @@ public class MCTSMomentum_AI extends MCTSMove_After_Attack_AI{
 					maxChild.setAttackSource("");
 					maxChild.setAttackDest("");
 					maxChild.updateHash(lastNode);
-					calculateMaxChildren(maxChild);
 
 					maxChild.depth = lastNode.depth + 1;
 					if (maxChild.depth > maxTreeDepth) {
@@ -281,7 +275,6 @@ public class MCTSMomentum_AI extends MCTSMove_After_Attack_AI{
 					if (maxChild.getAttackDest().length() == 0) {
 						maxChild.setTreePhase(GameTreeNode.MANOEUVRE);
 					}
-					calculateMaxChildren(maxChild);
 
 					maxChild.depth = lastNode.depth + 1;
 					if (maxChild.depth > maxTreeDepth) {
@@ -398,7 +391,6 @@ public class MCTSMomentum_AI extends MCTSMove_After_Attack_AI{
 				newChild.setTreePhase(GameTreeNode.ATTACK);
 				newChild.Momentum = true;
 			}
-			calculateMaxChildren(newChild);
 			newChild.updateHash(lastNode);
 			
 			newChild.depth = lastNode.depth + 1;
@@ -452,7 +444,6 @@ public class MCTSMomentum_AI extends MCTSMove_After_Attack_AI{
 
 					newChild.setMoveReq(false);
 
-					calculateMaxChildren(newChild);
 					newChild.updateHash(lastNode);
 					
 					newChild.depth = lastNode.depth + 1;
@@ -473,7 +464,7 @@ public class MCTSMomentum_AI extends MCTSMove_After_Attack_AI{
 			if (lastNode.numberOfManoeuvreBranches == 0) {
 				lastNode.numberOfManoeuvreBranches = Math.min(
 						params.MCTSManBranchQualityFactor,
-						lastNode.maxChildren / 4);
+						lastNode.maxChildren() / 4);
 			}
 
 			if(lastNode.manSources == null) {
@@ -520,7 +511,7 @@ public class MCTSMomentum_AI extends MCTSMove_After_Attack_AI{
 				double maxRating = Double.NEGATIVE_INFINITY;
 				MCTSNode maxChild = null;
 
-				if (lastNode.maxChildren == 1) {
+				if (lastNode.maxChildren() == 1) {
 					maxChild = lastNode.clone();
 					maxChild.setTreePhase(GameTreeNode.RECRUIT);
 					maxChild.switchMaxPlayer();
@@ -529,7 +520,7 @@ public class MCTSMomentum_AI extends MCTSMove_After_Attack_AI{
 				}
 
 				// Fix search range
-				if ((lastNode.maxChildren - lastNode.getChildren().size()) < lastNode.numberOfManoeuvreBranches + 1) {
+				if ((lastNode.maxChildren() - lastNode.getChildren().size()) < lastNode.numberOfManoeuvreBranches + 1) {
 					lastNode.numberOfManoeuvreBranches--;
 				}
 
@@ -538,7 +529,7 @@ public class MCTSMomentum_AI extends MCTSMove_After_Attack_AI{
 				}
 
 				for (int i = 0; i < lastNode.numberOfManoeuvreBranches; i++) {
-					int index = r.nextInt(lastNode.maxChildren);
+					int index = r.nextInt(lastNode.maxChildren());
 
 					int nrTroops = -1;
 					MCTSNode temp = null;
@@ -596,7 +587,6 @@ public class MCTSMomentum_AI extends MCTSMove_After_Attack_AI{
 						maxChild.switchMaxPlayer();
 						maxChild.getGame().changeCurrentPlayer();
 
-						calculateMaxChildren(maxChild);
 						maxChild.updateHash(lastNode);
 							
 						maxChild.depth = lastNode.depth + 1;
@@ -652,8 +642,6 @@ public class MCTSMomentum_AI extends MCTSMove_After_Attack_AI{
 					maxChild.setTreePhase(GameTreeNode.RECRUIT);
 
 					maxChild.setMoveReq(false);
-
-					calculateMaxChildren(maxChild);
 
 					maxChild.depth = lastNode.depth + 1;
 					if (maxChild.depth > maxTreeDepth) {
