+++ src/main/java/risk/aiplayers/paperplayers/S_MCTS_Baseline_AI.java
@@ -1,4 +1,4 @@
-package risk.paperplayers;
+package risk.aiplayers.paperplayers;
 
 import java.util.ArrayList;
 import java.util.Collection;
@@ -14,7 +14,7 @@ import risk.aiplayers.util.GameTreeNode;
 import risk.aiplayers.util.MCTSNode;
 import risk.commonObjects.Territory;
 
-public class S_MCTS_Naive_AI extends MonteCarloTreeSearchPlayer {
+public class S_MCTS_Baseline_AI extends MonteCarloTreeSearchPlayer {
 
 	MCTSNode globalNode;
 
@@ -23,14 +23,14 @@ public class S_MCTS_Naive_AI extends MonteCarloTreeSearchPlayer {
 	public static void main(String[] args) {
 		String tempName = args[0];
 		long time = Long.parseLong(args[1]);
-		new S_MCTS_AI(tempName, null, null, 2, time);
+			new S_MCTS_AI(tempName, null, null, 2, time);
 	}
 
-	public S_MCTS_Naive_AI(String name, String opp, String map, int id, long time) {
+	public S_MCTS_Baseline_AI(String name, String opp, String map, int id,
+			long time) {
 		super(name, opp, map, id, time, new AIParameter());
 	}
 
-	@Override
 	protected void calculateMaxChildren(MCTSNode lastNode) {
 		int count = 0;
 		switch (lastNode.getTreePhase()) {
@@ -119,7 +119,7 @@ public class S_MCTS_Naive_AI extends MonteCarloTreeSearchPlayer {
 			while (it.hasNext()) {
 				Territory t = it.next();
 				lastNode.getConnComponentBuckets().get(t.connectedRegion)
-				.add(t);
+						.add(t);
 			}
 
 			for (LinkedList<Territory> bucket : lastNode
@@ -142,7 +142,7 @@ public class S_MCTS_Naive_AI extends MonteCarloTreeSearchPlayer {
 		}
 		}
 
-		lastNode.maxChildren = count;
+		lastNode.setMaxChildren(count);
 	}
 
 	// EXPANSION (One child at a time)
@@ -168,36 +168,69 @@ public class S_MCTS_Naive_AI extends MonteCarloTreeSearchPlayer {
 			// System.out.println("Recruit 2");
 			MCTSNode maxChild = null;
 			double maxRating = Double.NEGATIVE_INFINITY;
-			// int count = 0;
-			while (maxChild == null) {
-				// count++;
-				// BO1
-				for (int i = 0; i < 1; i++) {
-					AIUtil.shuffleArray(perm);
 
-					MCTSNode tempChild = lastNode.clone();
+			if (lastNode.getChildren().size() == 0) {
+				// int maxID = 0;
+				String maxName = "";
+				double max = Double.MIN_VALUE;
 
-					Iterator<Territory> it = tempChild.getGame()
-							.getCurrentPlayer().getTerritories().values()
-							.iterator();
-					Territory current = it.next();
-					// Place troops according to permutation
-					for (int p = 0; p < perm.length; p++) {
-						if (perm[p]) {
-							current.incrementTroops();
-							tempChild.setAttackSource(current.getName());
-						} else {
-							current = it.next();
+				Iterator<Territory> it = lastNode.getGame().getCurrentPlayer()
+						.getTerritories().values().iterator();
+				while (it.hasNext()) {
+					Territory t = it.next();
+					double sum = 0.0;
+					for (Territory tn : t.getNeighbours()) {
+						if (lastNode.getGame().getOtherPlayer()
+								.getTerritoryByName(tn.getName()) != null) {
+							sum += tn.getNrTroops();
 						}
 					}
-					tempChild.setTreePhase(GameTreeNode.ATTACK);
-					double value = getValue(tempChild, lastNode);
-					if (value >= maxRating) {
-						maxRating = value;
-						maxChild = tempChild;
+					if (sum == 0.0)
+						continue;
+					double ratio = t.getNrTroops() / sum;
+					if (ratio > max) {
+						max = ratio;
+						// maxID = t.getId();
+						maxName = t.getName();
 					}
 				}
+				maxChild = lastNode.clone();
 
+				Territory source = maxChild.getGame().getCurrentPlayer()
+						.getTerritoryByName(maxName);
+				source.setNrTroops(source.getNrTroops() + n);
+				maxChild.setAttackSource(source.getName());
+				maxChild.setTreePhase(GameTreeNode.ATTACK);
+			} else {
+				// int count = 0;
+				while (maxChild == null) {
+					// count++;
+					for (int i = 0; i < params.MCTSRecruitBranchQualityFactor; i++) {
+						AIUtil.shuffleArray(perm);
+
+						MCTSNode tempChild = lastNode.clone();
+						tempChild.setTreePhase(GameTreeNode.ATTACK);
+						Iterator<Territory> it = tempChild.getGame()
+								.getCurrentPlayer().getTerritories().values()
+								.iterator();
+						Territory current = it.next();
+						// Place troops according to permutation
+						for (int p = 0; p < perm.length; p++) {
+							if (perm[p]) {
+								current.incrementTroops();
+								tempChild.setAttackSource(current.getName());
+							} else {
+								current = it.next();
+							}
+						}
+						double value = getValue(tempChild, lastNode);
+						if (value >= maxRating) {
+							maxRating = value;
+							maxChild = tempChild;
+						}
+					}
+
+				}
 			}
 
 			maxChild.setVisitCount(0);
@@ -208,7 +241,8 @@ public class S_MCTS_Naive_AI extends MonteCarloTreeSearchPlayer {
 			maxChild.setMoveReq(false);
 
 			calculateMaxChildren(maxChild);
-
+			maxChild.updateHash(lastNode);
+			
 			maxChild.depth = lastNode.depth + 1;
 			if (maxChild.depth > maxTreeDepth) {
 				maxTreeDepth = maxChild.depth;
@@ -222,10 +256,10 @@ public class S_MCTS_Naive_AI extends MonteCarloTreeSearchPlayer {
 			if (lastNode.numberOfAttackBranches == 0) {
 				lastNode.numberOfAttackBranches = Math.min(
 						params.MCTSAttackBranchQualityFactor,
-						lastNode.maxChildren / 2);
+						lastNode.maxChildren() / 2);
 			}
 
-			// Generate all and pick randomly
+			// Generate all and pick randomly from top 30
 
 			if (lastNode.attackChildren == null) {
 				lastNode.attackChildren = new ArrayList<MCTSNode>();
@@ -239,6 +273,9 @@ public class S_MCTS_Naive_AI extends MonteCarloTreeSearchPlayer {
 					noAttackChild.setValue(getValue(noAttackChild, lastNode));
 					// Add option to not attack
 					lastNode.attackChildren.add(noAttackChild);
+					
+					// Decrease Max Children then
+					lastNode.setMaxChildren(lastNode.maxChildren()-1);
 				}
 
 				// Populate treeset
@@ -280,41 +317,107 @@ public class S_MCTS_Naive_AI extends MonteCarloTreeSearchPlayer {
 			int count = 0;
 			while (true) {
 				count++;
+
+				// Fix search range
+				if (lastNode.attackChildren.size() < lastNode.numberOfAttackBranches + 1) {
+					lastNode.numberOfAttackBranches--;
+				}
+
+				if (lastNode.numberOfAttackBranches <= 1) {
+					lastNode.numberOfAttackBranches = lastNode.maxChildren() - lastNode.getChildren().size();
+				}
+				
+
 				MCTSNode maxChild = null;
 				double maxRating = Double.NEGATIVE_INFINITY;
 
-				for (int i = 0; i < 1; i++) {
-					index = r.nextInt(lastNode.attackChildren.size());
+				// BASELINE
+				if (lastNode.getChildren().size() == 0) {
 
-					MCTSNode temp = lastNode.attackChildren.get(index);
-					double value = temp.getValue();
-
-					if (value >= maxRating) {
-						maxRating = value;
-						maxChild = temp;
+					maxChild = lastNode.clone();
+					int max = Integer.MIN_VALUE;
+					int maxId = -1;
+					String maxName = "";
+					Iterator<Territory> it = lastNode.getGame()
+							.getCurrentPlayer().getTerritories().values()
+							.iterator();
+					while (it.hasNext()) {
+						Territory t = it.next();
+						if (t.getNrTroops() > max
+								&& !AIUtil.isHinterland(lastNode, t)) {
+							max = t.getNrTroops();
+							maxId = t.getId();
+							maxName = t.getName();
+						}
 					}
+					Territory attackSource = lastNode.getGame()
+							.getCurrentPlayer().getTerritoryByID(maxId);
 
-					// End game
-					if (count == 50 && value >= Double.MAX_VALUE - 1) {
-						maxChild.setVisitCount(0);
-						maxChild.setWinCount(0);
-						maxChild.setParent(lastNode);
-						maxChild.setChildren(new ArrayList<MCTSNode>());
+					if (attackSource.getNrTroops() > 1) {
+						int min = Integer.MAX_VALUE;
+						int minId = -1;
+						String minName = "";
+						for (Territory t : attackSource.getNeighbours()) {
+							if (t.getNrTroops() < min
+									&& lastNode.getGame().getOtherPlayer()
+											.getTerritoryByName(t.getName()) != null) {
+								min = t.getNrTroops();
+								minId = t.getId();
+								minName = t.getName();
+							}
+						}
 
-						if (maxChild.getAttackDest().length() == 0) {
-							lastNode.noAttackAdded = true;
+						if (minId != -1) {
+							maxChild.setAttackSource(maxName);
+							maxChild.setAttackDest(minName);
+							maxChild.setTreePhase(GameTreeNode.RANDOMEVENT);
+							maxChild.setValue(getWeightedEval(maxChild, lastNode));
+						} else {
+							maxChild.setAttackSource("");
+							maxChild.setAttackDest("");
 							maxChild.setTreePhase(GameTreeNode.MANOEUVRE);
+							maxChild.setValue(getValue(maxChild, lastNode));
 						}
-						calculateMaxChildren(maxChild);
-						maxChild.updateHash(lastNode);
+					} else {
+						maxChild = lastNode.clone();
+						maxChild.setAttackSource("");
+						maxChild.setAttackDest("");
+						maxChild.setTreePhase(GameTreeNode.MANOEUVRE);
+						maxChild.setValue(getValue(maxChild, lastNode));
+					}
+				} else {
+					for (int i = 0; i < lastNode.numberOfAttackBranches; i++) {
+						index = r.nextInt(lastNode.attackChildren.size());
 
-						maxChild.depth = lastNode.depth + 1;
-						if (maxChild.depth > maxTreeDepth) {
-							maxTreeDepth = maxChild.depth;
+						MCTSNode temp = lastNode.attackChildren.get(index);
+						double value = temp.getValue();
+
+						if (value >= maxRating) {
+							maxRating = value;
+							maxChild = temp;
 						}
 
-						lastNode.addChild(maxChild);
-						return maxChild;
+						// End game
+						if (count >= 50 && value >= Double.MAX_VALUE - 1) {
+							maxChild.setVisitCount(0);
+							maxChild.setWinCount(0);
+							maxChild.setParent(lastNode);
+							maxChild.setChildren(new ArrayList<MCTSNode>());
+
+							if (maxChild.getAttackDest().length() == 0) {
+								lastNode.noAttackAdded = true;
+								maxChild.setTreePhase(GameTreeNode.MANOEUVRE);
+							}
+							calculateMaxChildren(maxChild);
+
+							maxChild.depth = lastNode.depth + 1;
+							if (maxChild.depth > maxTreeDepth) {
+								maxTreeDepth = maxChild.depth;
+							}
+
+							lastNode.addChild(maxChild);
+							return maxChild;
+						}
 					}
 				}
 
@@ -341,6 +444,18 @@ public class S_MCTS_Naive_AI extends MonteCarloTreeSearchPlayer {
 						break;
 					}
 				}
+				
+				// Catching forever loop
+				if (lastNode.numberOfAttackBranches <= 1 && (lastNode.maxChildren() - lastNode.getChildren().size()) == 1) {
+					for (MCTSNode child : lastNode.getChildren()) {
+						if (child.getAttackSource().equalsIgnoreCase(
+								maxChild.getAttackSource())
+								&& child.getAttackDest().equalsIgnoreCase(
+										maxChild.getAttackDest())) {
+							return child;
+						}
+					}
+				}
 
 				// Add unique child to existing children
 				if (unique) {
@@ -365,6 +480,7 @@ public class S_MCTS_Naive_AI extends MonteCarloTreeSearchPlayer {
 					return maxChild;
 				}
 			}
+
 		}
 
 		case GameTreeNode.RANDOMEVENT: {
@@ -467,7 +583,7 @@ public class S_MCTS_Naive_AI extends MonteCarloTreeSearchPlayer {
 			}
 			calculateMaxChildren(newChild);
 			newChild.updateHash(lastNode);
-
+			
 			newChild.depth = lastNode.depth + 1;
 			if (newChild.depth > maxTreeDepth) {
 				maxTreeDepth = newChild.depth;
@@ -506,10 +622,10 @@ public class S_MCTS_Naive_AI extends MonteCarloTreeSearchPlayer {
 
 					AIUtil.resolveMoveAction(
 							newChild.getGame()
-							.getCurrentPlayer()
-							.getTerritoryByName(
-									newChild.getAttackSource()),
-									newChild.getGame()
+									.getCurrentPlayer()
+									.getTerritoryByName(
+											newChild.getAttackSource()),
+							newChild.getGame()
 									.getCurrentPlayer()
 									.getTerritoryByName(
 											newChild.getAttackDest()), troops);
@@ -520,7 +636,7 @@ public class S_MCTS_Naive_AI extends MonteCarloTreeSearchPlayer {
 
 					calculateMaxChildren(newChild);
 					newChild.updateHash(lastNode);
-
+					
 					newChild.depth = lastNode.depth + 1;
 					if (newChild.depth > maxTreeDepth) {
 						maxTreeDepth = newChild.depth;
@@ -529,14 +645,17 @@ public class S_MCTS_Naive_AI extends MonteCarloTreeSearchPlayer {
 					lastNode.addChild(newChild);
 					return newChild;
 				}
+
 			}
+
 		}
+
 		case GameTreeNode.MANOEUVRE: {
 
 			if (lastNode.numberOfManoeuvreBranches == 0) {
 				lastNode.numberOfManoeuvreBranches = Math.min(
 						params.MCTSManBranchQualityFactor,
-						lastNode.maxChildren / 4);
+						lastNode.maxChildren() / 4);
 			}
 
 			// System.out.println("Man 1");
@@ -568,9 +687,9 @@ public class S_MCTS_Naive_AI extends MonteCarloTreeSearchPlayer {
 										newChild.setManDest(dest);
 										newChild.setTreePhase(GameTreeNode.RECRUIT);
 										lastNode.manTroopBins
-										.add(lastNode.manTroopBins
-												.get(lastNode.manTroopBins
-														.size() - 1)
+												.add(lastNode.manTroopBins
+														.get(lastNode.manTroopBins
+																.size() - 1)
 														+ src.getNrTroops() - 1);
 										lastNode.manChildren.add(newChild);
 									}
@@ -579,6 +698,7 @@ public class S_MCTS_Naive_AI extends MonteCarloTreeSearchPlayer {
 						}
 					}
 				}
+
 			}
 
 			Random r = new Random();
@@ -588,96 +708,173 @@ public class S_MCTS_Naive_AI extends MonteCarloTreeSearchPlayer {
 				double maxRating = Double.NEGATIVE_INFINITY;
 				MCTSNode maxChild = null;
 
-				if (lastNode.maxChildren == 1) {
+				if (lastNode.maxChildren() == 1) {
 					maxChild = lastNode.clone();
 					maxChild.setTreePhase(GameTreeNode.RECRUIT);
 					maxChild.switchMaxPlayer();
 					maxChild.getGame().changeCurrentPlayer();
 					maxChild.updateHash(lastNode);
-				}				
+				}
 
-				for (int i = 0; i < 1; i++) {
-					int index = r.nextInt(lastNode.maxChildren);
+				// Fix search range
+				if ((lastNode.maxChildren() - lastNode.getChildren().size()) < lastNode.numberOfManoeuvreBranches + 1) {
+					lastNode.numberOfManoeuvreBranches--;
+				}
 
-					int first, last, middle = -1, nrTroops = -1;
-					MCTSNode temp = null;
+				if (lastNode.numberOfManoeuvreBranches < 1) {
+					lastNode.numberOfManoeuvreBranches = 1;
+				}
 
-					if (index > 0) {
-						first = 0;
-						last = lastNode.manTroopBins.size() - 1;
-						middle = (first + last) / 2;
+				// BASELINE
+				if (lastNode.getChildren().size() == 0) {
+					String minID = "";
+					int min = Integer.MAX_VALUE;
+					String maxID = "";
+					int max = Integer.MIN_VALUE;
+					
+					maxChild = lastNode.clone();
 
-						while (first <= last) {
-							int value = lastNode.manTroopBins.get(middle);
+					Iterator<Territory> it = maxChild.getGame()
+							.getCurrentPlayer().getTerritories().values()
+							.iterator();
+					while (it.hasNext()) {
+						Territory t = it.next();
+						if (t.getNrTroops() > max) {
+							max = t.getNrTroops();
+							maxID = t.getName();
+						}
+					}
+					Territory source = maxChild.getGame().getCurrentPlayer()
+							.getTerritoryByName(maxID);
 
-							if (value < index
-									&& index < lastNode.manTroopBins
-									.get(middle + 1)) {
-								nrTroops = index - value;
-								middle++;
-								break;
-							} else if (value == index) {
-								nrTroops = value
-										- lastNode.manTroopBins.get(middle - 1);
-								break;
-							} else if (value < index) {
-								first = middle + 1;
-							} else {
-								last = middle - 1;
-							}
+					it = maxChild.getGame().getCurrentPlayer().getTerritories()
+							.values().iterator();
+					while (it.hasNext()) {
+						Territory t = it.next();
+						if (t.getNrTroops() < min
+								&& t.connectedRegion == source.connectedRegion) {
+							min = t.getNrTroops();
+							minID = t.getName();
+						}
+					}
+
+					if (minID.length() != 0) {
+						Territory dest = maxChild.getGame().getCurrentPlayer()
+								.getTerritoryByName(minID);
+
+						if (source.getId() != dest.getId()) {
+							int total = source.getNrTroops()
+									+ dest.getNrTroops();
+							int nrTroops = (int) (total / 2.0);
+							maxChild.setManSource(source);
+							maxChild.setManDest(dest);
+							maxChild.setManTroopCount(nrTroops + "");
+
+							AIUtil.resolveMoveAction(
+									maxChild.getGame()
+											.getCurrentPlayer()
+											.getTerritoryByName(
+													maxChild.getManSource()
+															.getName()),
+									maxChild.getGame()
+											.getCurrentPlayer()
+											.getTerritoryByName(
+													maxChild.getManDest()
+															.getName()),
+									nrTroops);
+							maxChild.setTreePhase(GameTreeNode.RECRUIT);
+							maxChild.switchMaxPlayer();
+							maxChild.getGame().changeCurrentPlayer();
+							maxChild.updateHash(lastNode);
+						}
+					}
+				} else {
+					for (int i = 0; i < lastNode.numberOfManoeuvreBranches; i++) {
+						int index = r.nextInt(lastNode.maxChildren());
+
+						int first, last, middle = -1, nrTroops = -1;
+						MCTSNode temp = null;
+
+						if (index > 0) {
+							first = 0;
+							last = lastNode.manTroopBins.size() - 1;
 							middle = (first + last) / 2;
 
+							while (first <= last) {
+								int value = lastNode.manTroopBins.get(middle);
+
+								if (value < index
+										&& index < lastNode.manTroopBins
+												.get(middle + 1)) {
+									nrTroops = index - value;
+									middle++;
+									break;
+								} else if (value == index) {
+									nrTroops = value
+											- lastNode.manTroopBins
+													.get(middle - 1);
+									break;
+								} else if (value < index) {
+									first = middle + 1;
+								} else {
+									last = middle - 1;
+								}
+								middle = (first + last) / 2;
+
+							}
+							temp = lastNode.manChildren.get(middle).clone();
+
+							temp.setManTroopCount(nrTroops + "");
+							AIUtil.resolveMoveAction(
+									temp.getGame()
+											.getCurrentPlayer()
+											.getTerritoryByName(
+													temp.getManSource()
+															.getName()),
+									temp.getGame()
+											.getCurrentPlayer()
+											.getTerritoryByName(
+													temp.getManDest().getName()),
+									nrTroops);
+							temp.switchMaxPlayer();
+							temp.getGame().changeCurrentPlayer();
+
+						} else {
+							temp = lastNode.manChildren.get(0);
 						}
-						temp = lastNode.manChildren.get(middle).clone();
 
-						temp.setManTroopCount(nrTroops + "");
-						AIUtil.resolveMoveAction(
-								temp.getGame()
-								.getCurrentPlayer()
-								.getTerritoryByName(
-										temp.getManSource().getName()),
-										temp.getGame()
-										.getCurrentPlayer()
-										.getTerritoryByName(
-												temp.getManDest().getName()),
-												nrTroops);
-						temp.switchMaxPlayer();
-						temp.getGame().changeCurrentPlayer();
-					} else {
-						temp = lastNode.manChildren.get(0);
-					}
+						double value = getValue(temp, lastNode);
+						
+						// End game
+						if (count == 50 && value >= Double.MAX_VALUE - 1) {
+							maxChild = lastNode.clone();
+							maxChild.setVisitCount(0);
+							maxChild.setWinCount(0);
+							maxChild.setParent(lastNode);
+							maxChild.setChildren(new ArrayList<MCTSNode>());
 
-					double value = getValue(temp, lastNode);
+							maxChild.setTreePhase(GameTreeNode.RECRUIT);
 
-					// End game
-					if (count == 50 && value >= Double.MAX_VALUE - 1) {
-						maxChild = lastNode.clone();
-						maxChild.setVisitCount(0);
-						maxChild.setWinCount(0);
-						maxChild.setParent(lastNode);
-						maxChild.setChildren(new ArrayList<MCTSNode>());
+							maxChild.setMoveReq(false);
+							maxChild.switchMaxPlayer();
+							maxChild.getGame().changeCurrentPlayer();
 
-						maxChild.setTreePhase(GameTreeNode.RECRUIT);
+							calculateMaxChildren(maxChild);
+							maxChild.updateHash(lastNode);
+							
+							maxChild.depth = lastNode.depth + 1;
+							if (maxChild.depth > maxTreeDepth) {
+								maxTreeDepth = maxChild.depth;
+							}
 
-						maxChild.setMoveReq(false);
-						maxChild.switchMaxPlayer();
-						maxChild.getGame().changeCurrentPlayer();
-
-						calculateMaxChildren(maxChild);
-						maxChild.updateHash(lastNode);
-
-						maxChild.depth = lastNode.depth + 1;
-						if (maxChild.depth > maxTreeDepth) {
-							maxTreeDepth = maxChild.depth;
+							lastNode.addChild(maxChild);
+							return maxChild;
 						}
 
-						lastNode.addChild(maxChild);
-						return maxChild;
-					}
-
-					if (value >= maxRating) {
-						maxRating = value;
-						maxChild = temp;
+						if (value >= maxRating) {
+							maxRating = value;
+							maxChild = temp;
+						}
 					}
 				}
 
@@ -695,15 +892,15 @@ public class S_MCTS_Naive_AI extends MonteCarloTreeSearchPlayer {
 					if (maxChild.getManSource() != null
 							&& maxChild.getManDest() != null
 							&& child.getManSource()
-							.getName()
-							.equalsIgnoreCase(
-									maxChild.getManSource().getName())
-									&& child.getManDest()
+									.getName()
+									.equalsIgnoreCase(
+											maxChild.getManSource().getName())
+							&& child.getManDest()
 									.getName()
 									.equalsIgnoreCase(
 											maxChild.getManDest().getName())
-											&& child.getManTroopCount().equalsIgnoreCase(
-													maxChild.getManTroopCount())) {
+							&& child.getManTroopCount().equalsIgnoreCase(
+									maxChild.getManTroopCount())) {
 						unique = false;
 						break;
 					}
@@ -729,8 +926,10 @@ public class S_MCTS_Naive_AI extends MonteCarloTreeSearchPlayer {
 					return maxChild;
 				}
 			}
+
 		}
 		}
+
 		return null;
 	}
 
@@ -755,7 +954,7 @@ public class S_MCTS_Naive_AI extends MonteCarloTreeSearchPlayer {
 
 		int n = numberOfTroops;
 		int m = game.getCurrentPlayer().getTerritories().size();
-		root.maxChildren = (int) (AIUtil.nCk(n + m - 1, n));
+		root.setMaxChildren((int) (AIUtil.nCk(n + m - 1, n)));
 
 		treeDepth = 0;
 		maxTreeDepth = Integer.MIN_VALUE;
@@ -766,7 +965,7 @@ public class S_MCTS_Naive_AI extends MonteCarloTreeSearchPlayer {
 		MCTSNode action = MCTSSearch(root);
 		// double time = (System.nanoTime() - startTime) / 1000000.0;
 
-		//printStats(root, time);
+		// printStats(root, time);
 
 		/*
 		 * 
@@ -809,7 +1008,7 @@ public class S_MCTS_Naive_AI extends MonteCarloTreeSearchPlayer {
 		MCTSNode action = MCTSSearch(globalNode);
 		// double time = (System.nanoTime() - startTime) / 1000000.0;
 
-		//printStats(globalNode, time);
+		// printStats(globalNode, time);
 
 		/*
 		 * if (tv != null) tv.close(); tv = new TreeView(root);
@@ -879,10 +1078,10 @@ public class S_MCTS_Naive_AI extends MonteCarloTreeSearchPlayer {
 			temp.setTreePhase(GameTreeNode.ATTACK);
 			AIUtil.resolveMoveAction(
 					temp.getGame().getCurrentPlayer()
-					.getTerritoryByName(lastAttackSource.getName()),
+							.getTerritoryByName(lastAttackSource.getName()),
 					temp.getGame()
-					.getCurrentPlayer()
-					.getTerritoryByName(lastAttackDestination.getName()),
+							.getCurrentPlayer()
+							.getTerritoryByName(lastAttackDestination.getName()),
 					i);
 			double value = getValue(temp, null);
 			if (value >= maxValue) {
@@ -909,13 +1108,13 @@ public class S_MCTS_Naive_AI extends MonteCarloTreeSearchPlayer {
 
 				AIUtil.resolveMoveAction(
 						globalNode.getGame().getCurrentPlayer()
-						.getTerritoryByName(lastAttackSource.getName()),
+								.getTerritoryByName(lastAttackSource.getName()),
 						globalNode
-						.getGame()
-						.getCurrentPlayer()
-						.getTerritoryByName(
-								lastAttackDestination.getName()),
-								troops);
+								.getGame()
+								.getCurrentPlayer()
+								.getTerritoryByName(
+										lastAttackDestination.getName()),
+						troops);
 
 				globalNode.setMaxPlayer(true);
 				globalNode.setVisitCount(0);
@@ -947,7 +1146,7 @@ public class S_MCTS_Naive_AI extends MonteCarloTreeSearchPlayer {
 		MCTSNode action = MCTSSearch(globalNode);
 		// double time = (System.nanoTime() - startTime) / 1000000.0;
 
-		//printStats(globalNode, time);
+		// printStats(globalNode, time);
 
 		/*
 		 * if (tv != null) tv.close(); tv = new TreeView(root);
